Crud terminado sin diseño

+++ app/Http/Controllers/EmpleadoController.php
@@ -16,7 +16,7 @@ class EmpleadoController extends Controller
     public function index()
     {
         //
-        $datos['empleados'] = Empleado::paginate(5);
+        $datos['empleados'] = Empleado::paginate(7);
         return view ('empleado.index', $datos);
     }
 
@@ -48,7 +48,8 @@ class EmpleadoController extends Controller
         }
         Empleado::insert($datosEmpleado);
 
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje', 'Empleado registrado exitosamente');
     }
 
     /**
@@ -112,7 +113,12 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('/empleado/');
+        $empleado = Empleado::findOrFail($id);
+        
+        if(Storage::delete('public/'.$empleado->Foto)){
+            
+            Empleado::destroy($id);
+        }
+        return redirect('empleado')->with('mensaje','Empleado eliminado de la base de datos');
     }
 }
